#24257 - Admin product - No field on the product form to continue selling if out of stock (#5)
Added a property to Product.cs which is a boolean, this renders as a checkbox on the product add/edit page.

Added an event listener in ProductService.cs which runs before the update call to the db is made, this first checks if the event listener actually needs to fire; i.e. has the ContinueSellingWithNoStock field actually updated, then pulls the products variants in, if the variant shares the same value, it skips, else it updates the variant to share the same value.

I've also added the todo notes into EventListener.cs as I believe this will definitely come up at a later point.